updates to how expression matrix is included in cluster init

+++ localhost.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
-# must be run within metanetworkSynapse
 key=$1
 address=$2
-branch=$( git rev-parse --abbrev-ref HEAD )
+repo=$3
+cwd=$( cd $(dirname $0) ; pwd -P )/
 
 # move over essentials
-scp -i $key ~/.synapseConfig centos@$address:~/
-ssh -i $key centos@$address "git clone -b $branch https://github.com/philerooski/metanetworkSynapse.git /shared/metanetworkSynapse/"
+if [ -e ~/.synapseConfig ]; then
+    scp -i $key ~/.synapseConfig centos@$address:~/
+else
+    echo "~/.synapseConfig not found, skipping file transfer."
+fi
+ssh -i $key centos@$address "git clone $3 /shared/metanetworkSynapse/"
+scp -i $key $cwd/config.sh centos@$address:/shared/metanetworkSynapse/
 
 # run setup script on remote
 ssh -n -f -i $key centos@$address "sh -c 'nohup sh /shared/metanetworkSynapse/remote.sh > /dev/null 2>&1 &'"
